+++ src/ConsulSharp/V1/ConsulClientV1.cs
@@ -1,5 +1,6 @@
 ﻿using ConsulSharp.Core;
 using ConsulSharp.V1.ACL;
+using ConsulSharp.V1.Event;
 using ConsulSharp.V1.KeyValue;
 using ConsulSharp.V1.Status;
 
@@ -10,11 +11,14 @@ namespace ConsulSharp.V1
         public ConsulClientV1(Polymath polymath)
         {
             ACL = new ACLProvider(polymath);
+            Event = new EventProvider(polymath);
             KeyValue = new KeyValueProvider(polymath);
             Status = new StatusProvider(polymath);
         }
 
         public IACL ACL { get; }
+
+        public IEvent Event { get; }
 
         public IKeyValue KeyValue { get; }