+++ CHANGELOG.md
@@ -5,3 +5,4 @@
  * All ACL APIs - Create, Update, Read and Delete Token.
  * All KeyValue APIs - Write, Read, Delete Value.
  * All Status APIs - Get leader and Raft peers.
+ * All Event APIs - Fire and list events.